requirements.txt
spotipy==2.23.0
yt-dlp==2024.3.10
requests==2.31.0
PyQt5==5.15.9
tqdm==4.66.1
python-dotenv==1.0.0